+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Add `deprecate.property(obj, prop, message)`
+
 0.1.0 / 2014-06-15
 ==================